outsorce first visit monte carlo mouse to its own python script
